backend/requirements_automation.txt
pyautogui==0.9.54
pymongo==4.6.1
python-dotenv==1.0.0
